Option to display all Arsenal stats as text values.
This is intended for powerusers who don't want to have to resort to the config browser for stats that are shown in ACE.
Tooltip is a disclaimer about stats maybe not being 100% perfect, and a couple of noted examples (such as weapon range actually being max zeroing distance).
Not too much effort put into this, so everything uses a "default" text function that allows some modification and adds a suffix.